+++ tests/e2e/interOp/pytest.ini
@@ -33,7 +33,7 @@ wifi-SSID-2g-Pwd=12345678
 Default-SSID-5gl-perfecto-b=Default-SSID-5gl-perfecto-b
 Default-SSID-2g-perfecto-b=Default-SSID-2g-perfecto-b
 
-Default-SSID-perfecto-b=Default-SSID-2g-perfecto-b
+Default-SSID-perfecto-b=Default-SSID-5gl-perfecto-b
 # iOS Device Parameters
 #testbed-name=nola-ext-03
 ##equipment-model=ecw5410
